Database evolution

+++ mapper/backend/sql_js_backend.js
@@ -56,13 +56,31 @@ class SqlJsMapBackend extends MapBackend {
 			this.db = new Database();
 		}
 
+		this.s_getVersionNumber = this.db.prepare("PRAGMA user_version");
+
+		let gotVersion = this.getVersionNumber();
+		const wantVersion = this.getBackendVersionNumber();
+
+		// No version yet, let's see if there are any tables or else this is a fresh DB.
+		if(gotVersion === 0) {
+			if(this.db.prepare("SELECT name FROM sqlite_master WHERE type = 'table' AND name = 'entity'").get({}).length === 0) {
+				this.db.run("PRAGMA user_version = " + wantVersion);
+			}
+		}
+
+		gotVersion = this.getVersionNumber();
+
+		if(gotVersion !== wantVersion) {
+			throw new Error("version number does not match (got " + gotVersion + ", wanted " + wantVersion + ")");
+		}
+
 		this.db.run("PRAGMA foreign_keys = ON");
 		this.db.run("PRAGMA recursive_triggers = ON");
 
 		this.db.run("CREATE TABLE IF NOT EXISTS entity (entityid INTEGER PRIMARY KEY, type TEXT, valid BOOLEAN)");
 
 		// Node table and trigger to delete the corresponding entity when a node is deleted.
-		this.db.run("CREATE TABLE IF NOT EXISTS node (entityid INT PRIMARY KEY, parentid INT, FOREIGN KEY (entityid) REFERENCES entity(entityid) ON DELETE CASCADE, FOREIGN KEY (parentid) REFERENCES node(entityid) ON DELETE CASCADE)");
+		this.db.run("CREATE TABLE IF NOT EXISTS node (entityid INT PRIMARY KEY, nodetype TEXT, parentid INT, FOREIGN KEY (entityid) REFERENCES entity(entityid) ON DELETE CASCADE, FOREIGN KEY (parentid) REFERENCES node(entityid) ON DELETE CASCADE)");
 
 		if(this.options.buildDatabase) {
 			this.db.run("CREATE TRIGGER IF NOT EXISTS r_nodedeleted AFTER DELETE ON node FOR EACH ROW BEGIN DELETE FROM entity WHERE entityid = OLD.entityid; END");
@@ -71,7 +89,6 @@ class SqlJsMapBackend extends MapBackend {
 		// Triggers to cascade invalidation
 		if(this.options.buildDatabase) {
 			this.db.run("CREATE TRIGGER IF NOT EXISTS r_nodeinvalidated_children AFTER UPDATE OF valid ON entity WHEN NEW.type = 'node' AND NEW.valid = false BEGIN UPDATE entity SET valid = FALSE WHERE entityid IN (SELECT entityid FROM node WHERE parentid = NEW.entityid); END");
-			this.db.run("CREATE TRIGGER IF NOT EXISTS r_nodeinvalidated_edges AFTER UPDATE OF valid ON entity WHEN NEW.type = 'node' AND NEW.valid = false BEGIN UPDATE entity SET valid = FALSE WHERE entityid IN (SELECT edgeid FROM edge WHERE nodeid = NEW.entityid); END");
 		}
 
 		// Similar to nodes, a edge's corresponding entity will be deleted via trigger as soon as the edge is deleted.
@@ -101,19 +118,19 @@ class SqlJsMapBackend extends MapBackend {
 		this.s_entityValid = this.db.prepare("SELECT entityid FROM entity WHERE entityid = $entityId AND valid = TRUE");
 
 		this.s_createEntity = this.db.prepare("INSERT INTO entity (type, valid) VALUES ($type, TRUE)");
-		this.s_createNode = this.db.prepare("INSERT INTO node (entityid, parentid) VALUES ($entityId, $parentId)");
+		this.s_createNode = this.db.prepare("INSERT INTO node (entityid, parentid, nodetype) VALUES ($entityId, $parentId, $nodeType)");
 		this.s_createConnection = this.db.prepare("INSERT INTO edge (edgeid, nodeid) VALUES ($edgeId, $nodeId)");
+
+		this.s_getNodeType = this.db.prepare("SELECT nodetype FROM node WHERE node.entityid = $nodeId");
 
 		this.s_getNodeParent = this.db.prepare("SELECT nodep.entityid AS parentid FROM node AS nodep INNER JOIN node AS nodec ON nodep.entityid = nodec.parentid INNER JOIN entity ON entity.entityid = nodep.entityid WHERE entity.valid = true AND nodec.entityid = $nodeId");
 		this.s_getNodeChildren = this.db.prepare("SELECT node.entityid FROM node INNER JOIN entity ON node.entityid = entity.entityid WHERE parentID = $nodeId AND entity.valid = true");
-		this.s_getNodeEdges = this.db.prepare("SELECT edgeid FROM edge INNER JOIN entity ON entity.entityid = edge.edgeid WHERE nodeid = $nodeId AND entity.valid = true");
-		this.s_getEdgeNodes = this.db.prepare("SELECT nodeid FROM edge INNER JOIN entity ON nodeid = entity.entityid WHERE edgeid = $edgeId AND entity.valid = true");
+		this.s_getNodeEdges = this.db.prepare("SELECT edge1.edgeid FROM edge edge1 INNER JOIN edge edge2 ON (edge1.edgeid = edge2.edgeid AND edge1.nodeid != edge2.nodeid) INNER JOIN entity entity1 ON entity1.entityid = edge1.edgeid INNER JOIN entity entity2 ON entity2.entityid = edge2.nodeid INNER JOIN entity nodeentity1 ON nodeentity1.entityid = edge1.nodeid INNER JOIN entity nodeentity2 ON nodeentity2.entityid = edge2.nodeid WHERE edge1.nodeid = $nodeId AND entity1.valid = true AND entity2.valid = true AND nodeentity1.valid = TRUE AND nodeentity2.valid = TRUE");
+		this.s_getEdgeNodes = this.db.prepare("SELECT nodeid FROM edge INNER JOIN entity ON nodeid = entity.entityid WHERE edgeid = $edgeId");
 
-		this.s_getEdgeBetween = this.db.prepare("SELECT edge1.edgeid AS edgeid FROM edge edge1 INNER JOIN edge edge2 ON (edge1.edgeid = edge2.edgeid AND edge1.nodeid != edge2.nodeid) WHERE edge1.nodeid = $nodeAId AND edge2.nodeid = $nodeBId");
+		this.s_getEdgeBetween = this.db.prepare("SELECT edge1.edgeid AS edgeid FROM edge edge1 INNER JOIN edge edge2 ON (edge1.edgeid = edge2.edgeid AND edge1.nodeid != edge2.nodeid) INNER JOIN entity WHERE edge1.edgeid = entity.entityid AND edge1.nodeid = $nodeAId AND edge2.nodeid = $nodeBId AND entity.valid = TRUE");
 
-		this.s_getNodesInArea = this.db.prepare("SELECT node.entityid FROM node INNER JOIN property ON node.entityid = property.entityid INNER JOIN entity ON node.entityid = entity.entityid WHERE entity.valid = TRUE AND property.property = 'center' AND property.x >= $ax AND property.x <= $bx AND property.y >= $ay AND property.y <= $by AND property.z >= $az AND property.z <= $bz");
-
-		this.s_getNodesTouchingArea = this.db.prepare("SELECT node.entityid FROM node INNER JOIN property ON node.entityid = property.entityid INNER JOIN entity ON node.entityid = entity.entityid INNER JOIN property AS radiusproperty ON node.entityid = radiusproperty.entityid WHERE entity.valid = TRUE AND property.property = 'center' AND radiusproperty.property = 'radius' AND property.x >= $ax - radiusproperty.v_number AND property.x <= $bx + radiusproperty.v_number AND property.y >= $ay - radiusproperty.v_number AND property.y <= $by + radiusproperty.v_number AND property.z >= $az - radiusproperty.v_number AND property.z <= $bz + radiusproperty.v_number");
+		this.s_getNodesTouchingArea = this.db.prepare("SELECT node.entityid FROM node INNER JOIN property ON node.entityid = property.entityid INNER JOIN entity ON node.entityid = entity.entityid INNER JOIN property AS radiusproperty ON node.entityid = radiusproperty.entityid WHERE entity.valid = TRUE AND property.property = 'center' AND radiusproperty.property = 'radius' AND radiusproperty.v_number >= $minRadius AND property.x >= $ax - radiusproperty.v_number AND property.x <= $bx + radiusproperty.v_number AND property.y >= $ay - radiusproperty.v_number AND property.y <= $by + radiusproperty.v_number AND property.z >= $az - radiusproperty.v_number AND property.z <= $bz + radiusproperty.v_number");
 
 		// Triggers & foreign key constraints will handle deleting everything else relating to the entity.
 		this.s_deleteEntity = this.db.prepare("DELETE FROM entity WHERE entityid = $entityId");
@@ -137,12 +154,13 @@ class SqlJsMapBackend extends MapBackend {
 
 		/** Create a node atomically.
 		 * @param parentId {number|null} The ID of the node's parent, or null if none.
+		 * @param nodeType {string} The base type of the node.
 		 * @returns {number} The ID of the new node.
 		 */
-		this.baseCreateNode = (parentId) => {
+		this.baseCreateNode = (parentId, nodeType) => {
 			this.db.run("BEGIN EXCLUSIVE TRANSACTION");
 			const id = this.baseCreateEntity("node");
-			this.s_createNode.run({$entityId: id, $parentId: parentId ? parentId : null});
+			this.s_createNode.run({$entityId: id, $parentId: parentId ? parentId : null, $nodeType: nodeType});
 			this.db.run("COMMIT");
 			return id;
 		};
@@ -163,6 +181,15 @@ class SqlJsMapBackend extends MapBackend {
 
 		this.loaded = true;
 		await this.hooks.call("loaded");
+	}
+
+	getBackendVersionNumber() {
+		return 2;
+	}
+
+	getVersionNumber() {
+		const row = this.s_getVersionNumber.get({});
+		return row[0];
 	}
 
 	async getData() {
@@ -204,14 +231,21 @@ class SqlJsMapBackend extends MapBackend {
 		return this.getEntityRef(this.baseCreateEntity(type));
 	}
 
-	async createNode(parentId) {
-		const nodeRef = this.getNodeRef(this.baseCreateNode(parentId));
+	async createNode(parentId, nodeType) {
+		const nodeRef = this.getNodeRef(this.baseCreateNode(parentId, nodeType));
 		await nodeRef.create();
 		return nodeRef;
 	}
 
 	async createEdge(nodeAId, nodeBId) {
-		return this.getEdgeRef(this.baseCreateEdge(nodeAId, nodeBId));
+		const edgeRef = this.getEdgeRef(this.baseCreateEdge(nodeAId, nodeBId));
+		await edgeRef.create();
+		return edgeRef;
+	}
+
+	async getNodeType(nodeId) {
+		const row = this.s_getNodeType.get({$nodeId: nodeId});
+		return (row.length > 0 && row[0]) ? row[0] : null;
 	}
 
 	async getNodeParent(nodeId) {
@@ -283,15 +317,8 @@ class SqlJsMapBackend extends MapBackend {
 		});
 	}
 
-	async * getNodesInArea(box) {
-		this.s_getNodesInArea.bind({$ax: box.a.x, $ay: box.a.y, $az: box.a.z, $bx: box.b.x, $by: box.b.y, $bz: box.b.z});
-		while(this.s_getNodesInArea.step()) {
-			yield this.getNodeRef(this.s_getNodesInArea.get()[0]);
-		}
-	}
-
-	async * getNodesTouchingArea(box) {
-		this.s_getNodesTouchingArea.bind({$ax: box.a.x, $ay: box.a.y, $az: box.a.z, $bx: box.b.x, $by: box.b.y, $bz: box.b.z});
+	async * getNodesTouchingArea(box, minRadius) {
+		this.s_getNodesTouchingArea.bind({$ax: box.a.x, $ay: box.a.y, $az: box.a.z, $bx: box.b.x, $by: box.b.y, $bz: box.b.z, $minRadius: minRadius});
 		while(this.s_getNodesTouchingArea.step()) {
 			yield this.getNodeRef(this.s_getNodesTouchingArea.get()[0]);
 		}
